Fix asset loading
Though everything looked good on the dev server, assets (screen.css)
were broken with the output of `hugo --minify` - i.e. in production.

The theme was using absURL to get these URL's - which seems like it
should work, but idk.

Partial revert of 9923337 - but maintains the ability to use the dev server